// frontend/src/modules/user/product-search/types/product.filter.options.ts
export type ProductFilterOptions = {
  keyword?: string;
  status?: string;
  genres?: string[];
  authors?: string[];
  publisher?: string;
  series?: string;
  minPrice?: number;
  maxPrice?: number;
  rating?: number;
  page?: number;
  size?: number;
  sort?: SortType;
  hasPromotion?: string;
};

export type SortType =
  | "priceAsc"
  | "priceDesc"
  | "soldCount"
  | "rating"
  | "newest"| "";

export const SORT_OPTIONS: {
  value: SortType;
  label: string;
}[] = [
  {
    value: "priceAsc",
    label: "Giá: Thấp đến Cao",
  },
  {
    value: "priceDesc",
    label: "Giá: Cao đến Thấp",
  },
  {
    value: "soldCount",
    label: "Bán Chạy",
  },
  {
    value: "rating",
    label: "Đánh Giá Cao",
  },
  {
    value: "newest",
    label: "Mới Nhất",
  },
    {
    value: "",
    label: "Tất cả",
  },
];